requirements.txt
autopep8==2.0.4
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.6
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.27
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.0
